Feature: Kempton joystick support.
'~' is the fire button.

+++ Speculator/Speculator.Core/IPortHandler.cs
@@ -2,6 +2,6 @@ namespace Speculator.Core;
 
 public interface IPortHandler
 {
-    byte In(int portAddress);
+    byte In(ushort portAddress);
     void Out(byte port, byte b);
 }
